backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery[redis]==5.4.0
redis==5.2.1
aiohttp==3.11.11
requests==2.32.3
httpx==0.28.1
kafka-python==2.0.2
python-socketio==5.11.2
aio-pika==9.4.3
